itoa a reparer a priori ...

+++ libft/libft.h
@@ -6,7 +6,7 @@
 /*   By: fgarault <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/04/07 18:19:35 by fgarault          #+#    #+#             */
-/*   Updated: 2019/07/22 20:48:41 by fanny            ###   ########.fr       */
+/*   Updated: 2019/07/23 15:17:01 by fgarault         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -65,7 +65,7 @@ char				*ft_strsub(char const *s, unsigned int start, size_t len);
 char				*ft_strjoin(char const *s1, char const *s2);
 char				*ft_strtrim(char const *s);
 char				**ft_strsplit(char const *s, char c);
-char				*ft_itoa(long long n);
+char				*ft_itoa(int n);
 char				ft_putchar(char c);
 void				ft_putstr(char const *s);
 void				ft_putendl(char const *s);
